backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.4.0
django-suit==0.2.26
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
toml==0.10.1
tqdm==4.48.2
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
toml==0.10.1
tqdm==4.48.2
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
